requirements.txt
analytics-python==1.2.5
boto3==1.3.1
botocore==1.4.34
click==6.6
docutils==0.12
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
MarkupSafe==0.23
nose==1.3.7
pg8000==1.10.6
psycopg2==2.6.1
py==1.4.31
pytest==2.9.2
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.13
testing.common.database==1.1.0
testing.postgresql==1.3.0
Werkzeug==0.11.10
